requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-openai
langserve
python-dotenv
openai
groq
langchain-groq
bitsandbytes